worker/requirements.txt
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
structlog>=24.1.0
sentry-sdk[fastapi]>=1.40.0
